requirements.txt
streamlit
openai>=1.0.0
faiss-cpu
pypdf
python-dotenv
tiktoken
scikit-learn
numpy
pandas